review: package timeserie renamed to timeseries, add TimeUnit
